requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
requests